change TLS context init to default TLS_{client,server}_method
due to the change in method, the context can asynchronously request
handshaking at any time, so we need to track the last state when
switching states. this is only a partial fix as the current code
expects to write the request with a single call to SSL_write, and
expects to read the response with a single call to SSL_read. a
future enhancement should record request/response buffer offsets.